Ticket Office
You are analyzing sales data from a ticket office. The ticket for an adult is  20,𝑤ℎ𝑖𝑙𝑒𝑡ℎ𝑒𝑡𝑖𝑐𝑘𝑒𝑡𝑓𝑜𝑟𝑎𝑐ℎ𝑖𝑙𝑑𝑢𝑛𝑑𝑒𝑟18𝑖𝑠 5. The data you are given is in a dictionary format, where the keys are the sold ticket numbers, and the values are the customer ages. For example, "123-08": 24 means that the ticket was bought a 24 year old. Your goal is to calculate how much more money the office would make if it would change the ticket discount age to the given input. So, your program needs to take an integer as input and output the percentage of revenue growth, if the discount was given to people under that age.